Extracts team-related sections from account overview into a new My Team page
